Refactored more of the project
- Simplified the HTML and CSS structure
- Modified the JSON field data
- Refactored alert CSS class to be a member of the event class
- Moved the sub tier type to be under the stream elements folder
- Renamed the 'streamElements' folder to be 'stream-elements'
