+++ src/models/events/raid-event.ts
@@ -4,6 +4,8 @@ import { StreamEvent } from './stream-event';
 import { StreamEventType } from './stream-event-type';
 
 export class RaidEvent extends StreamEvent {
+    public alertCssClass: string;
+
     public alertSound: string;
 
     public html: string;
@@ -20,6 +22,7 @@ export class RaidEvent extends StreamEvent {
         }
 
         this.html = this.generateHtml();
+        this.alertCssClass = 'raid-alert';
         this.alertSound = FieldStore.Get<string>(FieldKeys.RaidAlertSound);
     }
