+++ src/models/events/stream-event-type.ts
@@ -2,7 +2,6 @@ export enum StreamEventType {
     Cheer,
     Follow,
     GiftedSubscription,
-    Host,
     Raid,
     Subscription,
 }
