+++ src/models/events/stream-event.ts
@@ -1,6 +1,8 @@
 import { StreamEventType } from './stream-event-type';
 
 export abstract class StreamEvent {
+    public abstract alertCssClass: string;
+
     public abstract alertSound: string;
 
     public abstract html: string;
