+++ src/models/stream-elements/event-received-detail.ts
