+++ src/models/stream-elements/index.ts
@@ -3,4 +3,5 @@ export { EventReceivedDetail } from './event-received-detail';
 export { FieldData } from './field-data';
 export { SessionData } from './session-data';
 export { StreamElementsApi } from './stream-elements-api';
+export { SubTier } from './sub-tier';
 export { WidgetLoadDetail } from './widget-load-detail';
