+++ src/models/stream-elements/widget-load-detail.ts
