+++ src/utilities/alert-helper.ts
@@ -1,12 +1,4 @@
-import {
-    CheerEvent,
-    Slide,
-    SlideContent,
-    StreamEvent,
-    StreamEventType,
-    SubEvent,
-    SubTier,
-} from '@models';
+import { Slide, SlideContent, StreamEvent } from '@models';
 
 /**
  * A helper class for managing alerts
@@ -27,101 +19,8 @@ export class AlertHelper {
         slide.classList.add('slide');
         slide.appendChild(content);
 
-        const alertCssClass = this.lookupAlertCssClass(event);
-        slide.classList.add(alertCssClass);
+        slide.classList.add(event.alertCssClass);
 
         return slide;
-    }
-
-    /**
-     * Looks up a corresponding CSS class for an event
-     *
-     * @param event The alert event
-     * @returns A CSS class of the alert
-     */
-    private static lookupAlertCssClass(event: StreamEvent): string {
-        if (!event) {
-            throw new Error(`Parameter 'event' cannot be null or undefined.`);
-        }
-
-        let cssClass: string;
-        switch (event.type) {
-            case StreamEventType.Cheer:
-                cssClass = this.lookupCheerAlertCssClass((event as CheerEvent).amount);
-                break;
-            case StreamEventType.Follow:
-                cssClass = 'follow-alert';
-                break;
-            case StreamEventType.GiftedSubscription:
-                cssClass = 'gifted-sub-alert';
-                break;
-            case StreamEventType.Host:
-                cssClass = 'host-alert';
-                break;
-            case StreamEventType.Raid:
-                cssClass = 'raid-alert';
-                break;
-            case StreamEventType.Subscription:
-                cssClass = this.lookupSubAlertCssClass((event as SubEvent).tier);
-                break;
-            default:
-                break;
-        }
-
-        return cssClass;
-    }
-
-    /**
-     * Looks up a corresponding CSS class for a cheer event
-     *
-     * @param amount The amount of bits of the cheer
-     * @returns The CSS class of the alert
-     */
-    private static lookupCheerAlertCssClass(amount: number): string {
-        if (amount <= 0) {
-            throw new Error(`Parameter 'amount' cannot be less than or equal to 0.`);
-        }
-
-        if (amount < 100) {
-            return 'cheer-alert-tier-1';
-        } else if (amount < 1000) {
-            return 'cheer-alert-tier-2';
-        } else if (amount < 5000) {
-            return 'cheer-alert-tier-3';
-        } else if (amount < 10000) {
-            return 'cheer-alert-tier-4';
-        } else {
-            return 'cheer-alert-tier-5';
-        }
-    }
-
-    /**
-     *
-     * Looks up a corresponding CSS class for a sub event
-     *
-     * @param tier The tier of the sub
-     * @returns The CSS class of the alert
-     */
-    private static lookupSubAlertCssClass(tier: SubTier): string {
-        let css: string;
-        switch (tier) {
-            case '1000':
-                css = 'sub-alert';
-                break;
-            case '2000':
-                css = 'sub-alert';
-                break;
-            case '3000':
-                css = 'sub-alert';
-                break;
-            case 'prime':
-                css = 'prime-sub-alert';
-                break;
-            default:
-                css = 'sub-alert';
-                break;
-        }
-
-        return css;
     }
 }
