+++ src/widget.html
@@ -1,22 +1,9 @@
 <script src="https://kit.fontawesome.com/7e2a25f815.js" crossorigin="anonymous"></script>
-
 <link href="https://fonts.googleapis.com/css?family={{fontFamily}}:400,{{fontWeight}}" rel="stylesheet" />
 
 <div class="container flex {{feedOrientation}} align-center">
-    <!-- <div class="badge slide-frame">
-        <div class="slide">
-            <div class="badge-content">
-                <video autoplay loop>
-                    <source src="{{badgeVideo}}" type="video/webm">
-                </video>
-            </div>
-        </div>
-    </div> -->
-
     <div class="bar slide-frame">
         <div class="slide">
-            <div class="slide-content">
-            </div>
         </div>
     </div>
